movie/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-crispy-forms==1.9.0
django-embed-video==1.3.1
idna==2.9
pytz==2019.3
requests==2.23.0
sqlparse==0.3.0
urllib3==1.25.8
